Jenkins unearthed that trtri also needs to have A marked const for the previous change to compile in all configs.
This goes deeper down the stack than the previous change. I also want to
note that the trtri interface in rocBLAS previously did only say A[in],
which was practically true (as the overwriting code path was never
executed - hence deleted) but technically wrong. It does however either
way not *not* represent the actual trtri LAPACK reference interface where
A is [in,out]:
http://www.netlib.org/lapack/explore-html/de/da7/dtrsm_8f_source.html

Hence, reference trtri is supposed to be an in-place inversion and our
implementation may need to be changed again.


[ROCm/rocblas commit: e549b8cb45e27ca65ab1436d11a13bc726fded55]