Remove global locale change in d_vector
Setting LC_NUMERIC to "" changes the global locale from whatever the
current locale is (defaults to the C locale) to the user's preferred
locale for numeric values. This is a change to the global state of the
program, and requires synchronization as it would result in undefined
behaviour if any other thread used a print function at the same time.

The only integer formatting affected by this change is the thousands
seperator, which may be added in the user's preferred locale. While
this may be nice to have in an error message, it's easier to reason
about the behaviour of the program if you set the locale immediately
upon program startup and leave it unchanged for the remainder of the
program.

This setting also affects the radix point used in floating point
formatting, which may be ',' in European locales and '.' in Canada or
the US. As this is a global setting, that change can have unexpected
consequences elsewhere in the code (like where numbers are printed
for a text-based file format, rather than for human viewing).


[ROCm/rocSOLVER commit: 706222a6ddbebc352d0d9d8d59e5e71d9de16618]