Use standard library timer functions (#259)
Replace clock_gettime with std::chrono::steady_clock for portability.
When running the rocsolver-bench, I'm seeing a change of ~10% in the
reported time for the CPU implementation, but the GPU time remains
unchanged. It appears that the change in CPU time is an actual change
in the time taken to perform the benchmarked operation, and not a
result of measurement error. The CPU benchmarking time also varies
when I make unrelated changes, so I suspect that the benchmarking of
the CPU implementation is just not as robust as the GPU benchmarking.

[ROCm/rocSOLVER commit: 266a3159b55ec98f6abbe6ed9330cdb6b2088215]