Rename cblas_* functions to cpu_* (#482)
* Rename cblas_ prefix to cpu_

The function names for the BLAS / LAPACK wrappers were a bit misleading
as the actual cblas library has not been used in rocSOLVER since
77b82ca768b963a8b2a8f2fa063b5d0e8ad195d5.

This change also drops the explicit template type from the calls to
these functions. In nearly all cases, it was superfluous.

* Drop commented-out function implementations

The commented function calls being deleted in this commit were
implemented using the actual cblas library, so they would would
require changes to work after being uncommented. They might as well
just be created when we need them.

* Fix template type deduction

* Address code review comments

[ROCm/rocSOLVER commit: 63b7ab573966c772a5d4bc626f6b69c380b20350]