Improve error reporting for zero-tolerance test failures (#517)
The rocsolver tests report certain error conditions (like unexpected
info values) by adding a positive integer to the measured error. This
can make it difficult to understand what is wrong when reviewing the
logs for a failed set of tests (especially when there are multiple
properties that could all increment the error).

The Google Test EXPECT functions will check a condition, but allow
the test to continue. This means that an error can be recorded at
the place where it occurs, without losing data.

[ROCm/rocSOLVER commit: cac76478d3a2fb860b6fd477dc77c2ca786e7456]