Use proper include path for cholmod.h. (#364)
* Use proper include path for cholmod.h.

It appears that the `SuiteSparse::CHOLMOD` CMake library advertises its include directory as `$PREFIX/include/suitesparse`, making the proper way to include it `#include <cholmod.h>`. When installed in the system `/usr` prefix, the `suitesparse/cholmod.h` path works by accident. When using a custom built suitesparse at a custom location, this is not so.

I believe that this change is the correct way to do it for all configurations, based on how the SuiteSparse library is set up.

* Update suitesparse's find module scripts

hipSOLVER's find module scripts (for suitesparse) were using `find_path`
on name `suitesparse/cholmod.h` instead of `cholmod.h`, which allowed
hipSOLVER to `#include <suitesparse/cholmod.h>`.

This commit updates the aforementioned cmake scripts, to honour the way
`Suitesparse::CHOLMOD` advertises its include directories.

---------

Co-authored-by: Julio Machado Silva <161654951+jmachado-amd@users.noreply.github.com>

[ROCm/hipSOLVER commit: c35541f64d3f791883fcde69d55ce805c31c8927]