Add out-of-bounds memory checking for GPUMem buffers and sub-buffers (#3666)
This change adds detection for GPU buffer out-of-bounds memory accesses.
This detection relies on the fact that the GPU will raise a memory
fault if an access into GPU buffers crosses a 2 MiB alignment
boundary *and* the memory adjacent to the boundary is not allocated, i.e.,
if the gpu memory is aligned to a 2 MiB page size boundary and the
memory access is either before the start of the buffer or after the
end of the buffer.

To make the problem more likely to occur, the memory returned to the
client is either allocated at the start of the buffer (to catch issues
before the start of the buffer) or at the end of the buffer (to catch
issues at the end of the buffer).

There are two different detection mechanisms for detecting out-of-bounds GPU buffer
memory accesses:

1.  The first adds a new argument to MIOPenDriver, --gpubuffer_check , which can be used to make the GPUMem class allocate GPU buffers that are aligned to 2 MiB alignment boundaries. Specifying 1 for this option causes the client memory to be allocated at the start of the buffer (to detect OOBs before the buffer's start) and specifying 2 for this option causes the client memory to be aligned at the end of the buffer (to detect OOBs after the buffer's end). Note that this is currently enabled for convolutions and batchnorm, although it is simple to add it for other problems. (We might also want to consider turning this checking on by default since this only affects MIOpenDriver usage.)

2.  The second method detects OOBs memory accesses for sub-buffers. This can be enabled via the MIOPEN_DEBUG_CHECK_SUB_BUFFER_OOB_MEMORY_ACCESS environment variable, and uses a technique similar to that used for the GPUMem class (but adapted for sub-buffers).

[ROCm/MIOpen commit: 20ddfa7d44f3ecbaea95515f676e18f61015e416]