Add CMake HIP language support (#740)
* Add CMake HIP language support

Add USE_HIPCXX option to build using CMake's support for the
HIP language. This option requires CMake 3.21.3 or newer.

The rocsolver clients shouldn't need to be built as HIP, but they
require rocblas_is_complex, which is only provided by the rocblas
headers when compiling as HIP. In the future, this could be
corrected.

Note that set_source_files_properties only affects the targets declared
in the same CMake file as the call to set the properties. As such, it
is difficult to use together with target_sources. All uses of latter
have therefore been removed. The project declaration in the clients
directory has also been removed so that PROJECT_SOURCE_DIR can always
be used to refer to the root directory.

* Enable verbose build

* Replace clients PROJECT_BINARY_DIR

The project() call has been removed from clients/CMakeLists.txt, so the
only project() call is at the rocsolver root. The references to the
PROJECT_BINARY_DIR in that file can be replaced by
CMAKE_CURRENT_BINARY_DIR.

* Limit AMDGPU_TARGETS to when USE_HIPCXX is OFF

* Address code review comments