[rocRoller] Allow more tile sizes (#817)
Enable more tile sizes in rocRoller. Also made a few changes to reduce
sgpr usage (only reduces it by a small amount, but enough to get the new
GEMM test working).

Changes to enable 192X256 tile size:
1. Change `tileCeilDivide` to convert the tile size to 64bit integer
when not a power of 2. This keeps the current behavior for all power of
2 sizes, but allows other sizes to work.
2. Fix multiply operation to copy literal values to vgprs when literal
size is not allowed.

Changes to reduce sgpr usage:
1. Change order that `cleanArguments` finds arguments within the
coordinate graph. This modifiers the `Visitor` class, but
`cleanArguments` is the only section of code using this part of the
`Visitor` class.
2. Increase the default `minLaunchTimeExpressionComplexity` value.