Cleanup Expressions (#1177)
This PR can be read commit by commit. The first two commits are the
relevant ones.

# First commit:
- Make Expressions that include `PositionalArgument`s easier to work
with:
   - Added register and data types
     - This enables rendering such expressions to string.
- Added `positionalArgument` helper function to make code that creates
these expressions more readable
 - Cleaned up expression creation in `ConnectWorkgroups`
 - Added tests for new and existing functionality.

# Second commit
- Split up `Expression_evaluate.cpp` into multiple files to speed up
compilation

## Technical Details

The main challenge was the fact that we were previously calling directly
into the specializations of `OperationEvaluatorVisitor` from the
`operator()` overloads of `EvaluateVisitor`. I changed this to call into
a new function template `evaluateOp()` which is manually instantiated
for all needed types in different `.cpp` files.

> [!IMPORTANT]
> This means that when adding a new expression type, we may not
encounter a compiler error for a missing evaluate function. It will be a
linker error if the new expression type fits into one of the existing
subtypes.

After moving the relevant code into `_unary.cpp`, `_binary.cpp`, and
`_ternary.cpp`, `Expression_evaluate_binary.cpp` was still taking over 2
minutes to compile. This is because of the need to consider each pair of
types for each operation. We can reduce this by splitting out the
pointer types into their own sub-variant but that is a more involved
change. Instead I split the binary operations further into arithmetic,
comparison, logical/bitwise, and shift/SRConvert.

`Expression_evaluate_binary_arithmetic.hpp` is still the longest file to
build but it takes 1:16, down from 2:50.

The total time for a clean build of default targets is now 1:48 down
from 3:01.

# `DataTypes.hpp` refactor
 - Add `friendlyTypeName<T>()` utility function
- Make `DataTypes.hpp` more readable by moving much into new
`DataTypes_impl.hpp`
 - Get rid of a statically initialized variables in `DataTypes.cpp`

---------

Co-authored-by: Nathan Henderson <35206064+ThanHenderson@users.noreply.github.com>