Add llvm/lib to rpath (#1423)
To ensure libomp is searched for in opt/rocm/llvm/lib we add this path
relative to origin in addition to lib.

## Motivation

We want to use the rocm libomp. Adding `llvm/lib` to the install rpath
ensures that we find libomp relative to the executables such as
hipblaslt-bench and hipblaslt-test.

## Technical Details

Adds `$ORIGIN/../llvm/lib` to rpath. It is important to note that this
does not guarantee consistency in environments where there is an OpenMP
installation in `/usr` and in `/opt/rocm` which does occur in some of
our test environments. Another possibility is there is an install in
`/usr` at build time and then at runtime `/opt/rocm` (or the reverse).
CMake may then find the OpenMP in `/usr` and then at runtime the
executable uses the install in `/opt/rocm` and so on.

## Test Plan

CI pipelines and local build.