[rocThrust] Remove obsolete architectures from default list (#1441)
## Motivation

The size of our fat binary (as generated by building with all default
architectures enabled) is growing large enough that it is causing
linking issues for some of our libraries.

## Technical Details

This change removes gfx803 and gfx900 from the list of default
architectures to build for. This should help reduce the size of the fat
binary.

## Test Plan

Build rocThrust for all default architectures. Ensure there are no build
errors.

## Test Result

No build errors observed.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.