[rocSPARSE] sync before hipFree. (#1705)
Since HIP 7.0 `hipFree` is not implicitly blocking as before (see
https://rocm.docs.amd.com/projects/HIP/en/latest/hip-7-changes.html#update-hipfree
). Specifically if we allocate using `hipMallocAsync` and then try to
free using `hipFree`, the `hipFree` will not be blocking. This means
that in the cases where we allocate using `hipMallocAsync` and then free
using `hipFree` (for example when the stream is not available for the
free), we need to do an explicit synchronization.

This PR adds required synchronization points.

This PR also reverts #1609 as it is not required anymore.