rocfft: handle 4-D bricks in DebugPrintHash (#1897)
## Motivation

rocFFT aborts when LAYER=16 logging is enabled during multi-GPU
transforms.

## Technical Details

Batch dimension is treated by bricks as a 4th dimension, and these
bricks are directly passed to the logging code. So allow LAYER=16
logging to handle this case.

## Test Plan

Manual testing for this non-default developer feature.

## Test Result

Logging works successfully now in this situation.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.