Remove mutable autoGSU (#1944)
## Motivation

The mutable autoGSU member of the ContractionSolution class can cause a
race condition where kernels are launched with different autoGSU values
than they would otherwise select.

## Technical Details

This change removes the mutable autoGSU member variable and calculates
it as needed for a given problem+hardware pair to ensure the value is
accurate and consistent between solution selection and kernel launch.

## Test Plan

Tested with end-to-end application tests. A new unit test for this bug
will have to be done as follow-up work to allow the fix to be merged
quickly.

## Test Result

Works in initial application tests, awaiting results of additional
application runs.

## Submission Checklist

- [X] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.