Make max CUs setting work with dynamic grid prediction (#2260)
## Motivation

Reposting changes from PR #1950, updated to work with the latest
refactor changes in Origami.

For apps that overlap gemm and comm work, it is useful to limit the
number of CUs that can run gemm. But locking a gemm call to a specific
number of CUs can degrade performance due to poor work distribution.

This change updates the TENSILE_STREAMK_MAX_CUS setting so that it works
in combination with dynamic grid prediction, to select a good launch
grid within a requested CU limit.

## Technical Details

Changes dynamic grid prediction to accept an argument for max_cus. The
arg defaults to 0, meaning any number of CUs is fine. If it is set, the
default dynamic grid mode will work within the limitation to select a
launch grid that should perform well.

## Test Plan

Tested locally, env vars used for debugging do not currently have a
testing system.

## Test Result

Works as expected.

## Submission Checklist

- [X] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.