[rocRoller] Add swizzleTileSize support to GEMM client. (#2244)
## Motivation

The default swizzle-tile-size was changed recently and some performance
tests showed regressions. We didn't have a way of changing the
swizzle-tile-size. This PR adds that ability to the client.

## Technical Details

Adds an MKNLTuple class (MxK/NxL) to represent the swizzle tile size.
Adds a `--sts` option to the client (short for "swizzle tile size") that
parses a MxK/NxL tuple and uses this to set the swizzle-tile-size when
building the command.

While adding this MKNLTuple, I moved around some of the client argument
parsing code as well and added more tests.

## Test Plan

Adds a `test_gemm_options` test to `test_gemm_client.py` that tests the
command line parsing routines for `--mi`, `--wgts`, `--lds`, `--d2lds`,
`--mxlds`, `--mxd2lds`, and the new `--sts`.

Performance team has verified that the regression went away.