Set AutoWGM on By Default in Origami Libraries (#2386)
## Motivation

This PR enables setting WorkGroupMapping to 0 and enables picking a
default value (determined by the hardware specs) for it at runtime.
If it is combined with WGMXCCn1, Origami model predicts the best
combination for the two values (WGM and WGMXCC).
At the moment, Origami libs are modified to support it, while the other
libraries remain the same. Note that some libraries are modified to have
WGM1 instead of WGM0. These two are identical in terms of performance,
so there is no complication with that.

## Technical Details

With this PR and #2170:
- WGM0: default value for WGM is picked at runtime based on the hardware
- WGMXCCn1: default value for WGMXCC is picked at runtime based on the
hardware
- WGM0 and WGMXCCn1: Origami predicts the best combination.

## Test Plan

CI and locally. 

## Test Result

Improves the performance of Origami libs while not changing the
performance of the other libraries.

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.