[hipblas] export coverage report for upload to codecov.io (#2353)
## Motivation

We plan to upload code coverage reports to codecov.io

## Technical Details

Added a command to the `coverage` target to export the llvm-cov format
to lcov because that is the only format that codecov.io supports. CI
will handle the rest.

## Test Plan

A comment will be made on this PR with the coverage report from
codecov.io indicating that the process was successful