[Ex CI] Update Azure-Ci Dispatcher (#2457)
## Motivation
Currently, for non targeted branches are getting stuck on the Azure CI
Summary check.
<!-- Explain the purpose of this PR and the goals it aims to achieve.
-->

## Technical Details
This change checks to see what the target branch is for the PR, and if
it is one of the targeted branches then the azure ci will get
dispatched. Otherwise it will skip to the summary check and output an
empty summary. This should allow for the Azure CI Summary check to be
bypassed for branches that do not have Azure CI.
<!-- Explain the changes along with any relevant GitHub links. -->

## Test Plan

<!-- Explain any relevant testing done to verify this PR. -->

## Test Result

<!-- Briefly summarize test outcomes. -->

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.