[Github Actions] Added monorepo_source_of_truth flag (#2211)
## Motivation
Create another flag per project on repos-config.json to present the
super-repo being the source of truth, instead of the old individual
repo. This allows projects to not have this source of truth designation
tied to auto-subtree-push, as we may not want that behaviour for all
projects on the super repo and eventually want to turn off this feature.
<!-- Explain the purpose of this PR and the goals it aims to achieve.
-->

## Technical Details

<!-- Explain the changes along with any relevant GitHub links. -->

## Test Plan

<!-- Explain any relevant testing done to verify this PR. -->

## Test Result

<!-- Briefly summarize test outcomes. -->

## Submission Checklist

- [ ] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.