Update Db file locking for NFS (#1612)
## Motivation

In the event that MIOpen is used with a shared filesystem for user dbs,
it will need the lock files to be accessible to each node updating the
db files.
This stores lockfiles relative to the userdb directory so they can be
accessed by network filesystem.
The NFS is accessed by different nodes with separate OS, so the locking
needs to be done with shared files.

## Technical Details

Updates file paths to shared directory.
New lockfiles added, discrete to each hostname.
Shared lock secured by use of filesystem link function and lock claim
verified by link count.

## Tests

Verified using a list of 370 configurations on vultr cluster using NFS. 
Same set ran in 8 processes on 4 nodes for 32 competing processes.