fix(rocthrust): fix compilation error when compiling with gnu++17 standard (#2760)
## Motivation

In certain cases, compiling with `-std=gnu++17` would cause compile
errors.

## Technical Details

This change disables defining the sequential implementation whenever we
have the HIP runtime. The sequential implementation is not required in
that case, since if the HIP runtime is available we always want to run
the parallel implementation (i.e. we're on host!). When compile for
device, both parallel and sequential need to be defined. The sequential
one is required due to it the implementation being requested. The
parallel one is required because the device compiler needs to know what
kernels are launched from host. We therefore define it, but never call
it in our device compilation.

## Test Plan

Compiling the following program with `hipcc -std=gnu++17 main.cpp` must
work after installing rocThrust.

```cpp
// main.cpp
int main() {
    using K = __int128;
    using V = int;
    const int N = 4;
    thrust::device_vector<K> keys(N);
    thrust::device_vector<V> vals(N);
 
    K hkeys[N];
    for (int i = 0; i < N; ++i) hkeys[i] = (K)i * 100;
 
    thrust::copy(hkeys, hkeys + N, keys.begin());
    thrust::sort_by_key(keys.begin(), keys.end(), vals.begin()); // ---- Fails here ----
    return 0; 
}
```

## Test Result

Test plan succeeded.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.