[MIOpen] Provide initial set of 3D CK heuristics for MI355 (#2786)
## Motivation

Prior to this change, MIOpen did not have any heuristics for MI355 3D
shapes, which meant that customers that relied on out-of-the-box (OOTB)
performance might see a performance regression on some 3D shapes when
running on MI355.

## Technical Details

A previous change had added 3D CK-specific heuristics for MI300. In
order to extend the 3D CK heuristics coverage to MI355, this PR takes
the relevant heuristics and models and applies them to the MI355.

## Test Plan

Gather OOTB performance numbers for MI355 to compare the performance
with and without the new MI355 3D heuristics.

## Test Result

Performance testing has been done, which shows significant performance
for 3D CK fwd shapes. Note that the impact to bwd and wrw shapes was
negligible since the MIOpen WTI fallback did not choose the relevant CK
solver.

## Submission Checklist

- [x] Look over the contributing guidelines at
https://github.com/ROCm/ROCm/blob/develop/CONTRIBUTING.md#pull-requests.

---------

Co-authored-by: Christopher Erb <christopher.erb@amd.com>